Move components to right folders
