+++ frontend/app/movies/[id]/starRating.tsx
