+++ frontend/app/users/[id]/profileInfo.tsx
